requirements.txt
streamlit==1.28.1
langchain==0.0.354
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4
openai==1.6.1
python-dotenv==1.0.0
pypdf==6.0.0
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
httpx==0.25.2
sentence-transformers==2.2.2